pyproject.toml
[tool.poetry]
name = "audio-processing-api"
version = "1.0.0"
description = "High-performance audio processing service providing voice cloning and translation capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0"
celery = "^5.3.6"
redis = "^5.0.1"
boto3 = "^1.34.0"
loguru = "^0.7.2"
prometheus-client = "^0.19.0"
python-dotenv = "^1.0.0"
greenlet = "^2.0.1"

# AI/ML dependencies
torch = "^2.1.2"
torchaudio = "^2.1.2"
numpy = "^1.26.3"
scipy = "^1.12.0"
librosa = "^0.10.1"
soundfile = "^0.12.1"
ffmpeg-python = "^0.2.0"
yt-dlp = "^2024.1.1"

# Core ML packages
cython = "^3.0.8"
faster-whisper = "1.0.3"
ctranslate2 = "4.0.0"
tokenizers = "^0.15.1"
TTS = "0.22.0"

# Utility packages
python-magic = "^0.4.27"  # For file type detection

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"